Remove ocagent/cli.go and use viper for pprof setup
